Tidy up, moved files to correct directories
