+++ components/block/LatestBlocks.tsx
@@ -7,18 +7,18 @@ import TableContainer from "@material-ui/core/TableContainer";
 import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
 import Paper from "@material-ui/core/Paper";
-import Link from "../components/Link";
+import Link from "../Link";
 import Router from "next/router";
 import Grid from "@material-ui/core/Grid";
 import { Divider } from "@material-ui/core";
 import Typography from "@material-ui/core/Typography";
 import cx from "clsx";
 import Card from "@material-ui/core/Card";
-import Layout from "../components/Layout";
-import theme from "../themes/celo-theme";
+import Layout from "../Layout";
+import theme from "../../themes/celo-theme";
 import moment from "moment";
 import Hidden from "@material-ui/core/Hidden";
-import PriceCard from "./PriceCard";
+import PriceCard from "../PriceCard";
 import gql from "graphql-tag";
 import { useQuery } from "@apollo/react-hooks";
 import TablePagination from "@material-ui/core/TablePagination";
