+++ pages/blocks.tsx
@@ -4,7 +4,7 @@ import Layout from "../components/Layout";
 import ChartData from "../components/ChartData";
 import LatestTransactions from "../components/LatestTransactions";
 import Grid from "@material-ui/core/Grid";
-import LatestBlocks from "../components/LatestBlocks";
+import LatestBlocks from "../components/block/LatestBlocks";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
