+++ pages/transaction/[transaction].tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
 import Layout from "../../components/Layout";
 import Grid from "@material-ui/core/Grid";
-import TransactionDetails from "../../components/TransactionDetails";
+import TransactionDetails from "../../components/transaction/TransactionDetails";
 import { useRouter } from "next/router";
 
 const useStyles = makeStyles((theme: Theme) =>
